Reimplemented IO formats

+++ include/Types.hpp
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-using GridIndexType = int64_t;
+using GridIndexType = size_t;
 using FlatVectorIndexType = size_t;
 using StepType = int64_t;
